internal/netutil, pkg/schema: remove nocgo tag; os/user has its own nowadays
The os/user package nowadays says:

    When cgo is available, and the required routines are implemented
    in libc for a particular platform, cgo-based (libc-backed) code
    is used. This can be overridden by using osusergo build tag,
    which enforces the pure Go implementation.

So we don't need to work around it anymore.
