--- pkg/schema/lookup_nocgo.go
@@ -1,30 +1,0 @@
-//go:build nocgo
-
-/*
-Copyright 2016 The Perkeep Authors.
-
-Licensed under the Apache License, Version 2.0 (the "License");
-you may not use this file except in compliance with the License.
-You may obtain a copy of the License at
-
-     http://www.apache.org/licenses/LICENSE-2.0
-
-Unless required by applicable law or agreed to in writing, software
-distributed under the License is distributed on an "AS IS" BASIS,
-WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-See the License for the specific language governing permissions and
-limitations under the License.
-*/
-
-package schema
-
-func lookupUserToId(name string) (uid int, ok bool) {
-	return
-}
-
-// lookupMu is held
-func lookupUserid(id int) string {
-	parsedPasswd = true
-	populateMap(uidName, nil, "/etc/passwd")
-	return uidName[id]
-}
